Simplistic implementation of Karatsuba method relying on integers, standard sum and no overflows.

+++ src/test/java/org/vtsukur/algorithms/divide_and_conquer/KaratsubaMultiplicationTest.java
@@ -10,10 +10,10 @@ public class KaratsubaMultiplicationTest {
 
     @Test
     public void fourDigitMultiplication() {
-        final long first = 1234;
-        final long second = 5678;
+        final int first = 1234;
+        final int second = 5678;
 
-        final long product = new KaratsubaMultiplication().multiply(first, second, 4);
+        final int product = new KaratsubaMultiplication().multiply(first, second, 4);
 
         Assert.assertEquals(first * second, product);
     }
